Start getting rid of mapper and actor everywhere
One challenge is that we used the mappers to keep track of both
the dataset (mapper input) as well as lookuptable, so now we'll have
to do that explicitly.

Started in update_input, and while animation was mostly still working,
the tests involving masking were broken.

Left off in isofillpipeline...